requirements.txt
flask
torch
torchvision
pandas
numpy
pyarrow
opencv-python
pillow
